+++ IPCore/IPCore.h
@@ -410,7 +410,7 @@ uint8_t *arp_resolve(uint32_t node_ip_addr);
 void udp_recv(uint32_t from_addr, uint16_t from_port, uint16_t to_port, uint8_t * data, uint16_t data_len);
 //послать UDP пакет, возвращает 0 или 1
 //send UDP packet, returns 0 or 1
-uint8_t udp_send(uint32_t to_addr, uint32_t to_port, uint32_t from_port, uint8_t *data, uint16_t data_len);
+uint8_t udp_send(uint32_t to_addr, uint16_t to_port, uint16_t from_port, uint8_t *data, uint16_t data_len);
 
 
 // -------------------- DNS -------------------------------
